refactors providers
- Split `models.py` into `models/__init__.py` (shared types like `PlannedOutageEvent`, `YasnoRegion`) and `models/providers.py` (provider-specific classes)
- Moved `DTEK_PROVIDER_URLS` from models to `const.py` for better constant organization
- Updated imports across config flow, coordinators, and test files to reflect new structure
- Deleted the old `models.py` file
